+++ sample/src/main/kotlin/com/github/panpf/sketch/sample/ui/video/LocalVideoItemFactory.kt
@@ -20,11 +20,12 @@ import com.github.panpf.sketch.displayImage
 import com.github.panpf.sketch.request.updateDisplayImageOptions
 import com.github.panpf.sketch.request.videoFramePercent
 import com.github.panpf.sketch.sample.R
+import com.github.panpf.sketch.sample.appSettingsService
 import com.github.panpf.sketch.sample.databinding.VideoItemBinding
 import com.github.panpf.sketch.sample.model.VideoInfo
-import com.github.panpf.sketch.sample.appSettingsService
 import com.github.panpf.sketch.sample.ui.common.list.MyBindingItemFactory
 import com.github.panpf.sketch.stateimage.IconStateImage
+import com.github.panpf.sketch.stateimage.ResColor
 import com.github.panpf.sketch.stateimage.saveCellularTrafficError
 
 class LocalVideoItemFactory :
@@ -36,8 +37,9 @@ class LocalVideoItemFactory :
         item: BindingItem<VideoInfo, VideoItemBinding>
     ) {
         binding.videoItemIconImage.updateDisplayImageOptions {
-            placeholder(IconStateImage(R.drawable.ic_image_outline, R.color.placeholder_bg))
-            error(IconStateImage(R.drawable.ic_error, R.color.placeholder_bg)) {
+            val bg = ResColor(R.color.placeholder_bg)
+            placeholder(IconStateImage(R.drawable.ic_image_outline, bg))
+            error(IconStateImage(R.drawable.ic_error, bg)) {
                 saveCellularTrafficError(R.drawable.im_save_cellular_traffic)
             }
             crossfade()